Refactor vault tile
